fix: Player와 Dealer의 중복 코드 제거 (추상 클래스 이용)
